FaceRecognition/requirements.txt
numpy
matplotlib
pillow
opencv-contrib-python-3.7.0
pypyodbc
